Move the Plan module under src/shared
The Plan schema is imported by both sides — validateStateChange and the
Proposal tools on the server, the Plan Panel and the applier in the client — so
it is neither src/server nor src/client. It sat at src/plan, which meant each
tsconfig had to name it, and the next shared domain would have been appended to
both lists again.

src/shared is now the one root each tsconfig lists, and shared domains nest
inside it.
