+++ src/shared/plan/scope.ts
