+++ src/shared/plan/word-count.ts
