+++ tsconfig.client.json
@@ -18,5 +18,5 @@
     "composite": true,
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.client.tsbuildinfo"
   },
-  "include": ["src/plan", "src/client", ".storybook"]
+  "include": ["src/shared", "src/client", ".storybook"]
 }
