+++ tsconfig.worker.json
@@ -17,5 +17,5 @@
     "composite": true,
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.worker.tsbuildinfo"
   },
-  "include": ["src/plan", "src/server", "test", "worker-configuration.d.ts"]
+  "include": ["src/shared", "src/server", "test", "worker-configuration.d.ts"]
 }
